# docs/api-server/openapi.yaml
openapi: "3.1.0"
info:
  title: AgentClash API
  version: 1.0.0
  description: >
    B2B AI agent evaluation platform. Pits agent deployments against challenge
    packs, with replay inspection, scoring, comparison, and release gating.
  license:
    name: FSL-1.1-MIT
    url: https://github.com/agentclash/agentclash/blob/main/LICENSE
  contact:
    url: https://github.com/agentclash/agentclash
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /artifacts/{artifactID}/content:
    get:
      operationId: getArtifactContent
      tags: [Artifacts]
      summary: Download artifact content
      description: >
        Streams artifact binary content. Requires a valid HMAC-signed URL
        obtained from the authenticated download endpoint. A 30-second grace
        period is allowed past the expiry timestamp.
      security:
        - artifactSignature: []
      parameters:
        - $ref: "#/components/parameters/ArtifactID"
        - name: expires
          in: query
          required: true
          schema:
            type: integer
            format: int64
          description: Unix timestamp when the signed URL expires
        - name: signature
          in: query
          required: true
          schema:
            type: string
          description: HMAC-SHA256 signature
      responses:
        "200":
          description: Artifact binary content
          headers:
            Content-Type:
              schema:
                type: string
            Content-Disposition:
              schema:
                type: string
            Content-Security-Policy:
              schema:
                type: string
                example: "sandbox; default-src 'none'"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ---------------------------------------------------------------------------
  # Hosted integration
  # ---------------------------------------------------------------------------
  /v1/integrations/hosted-runs/{runID}/events:
    post:
      operationId: ingestHostedRunEvent
      tags: [HostedRuns]
      summary: Ingest hosted run event
      description: >
        Webhook-style endpoint for externally hosted agents to report execution
        events. Authenticated via an HMAC-signed callback token (not JWT).
      security:
        - hostedRunToken: []
      parameters:
        - $ref: "#/components/parameters/RunID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HostedRunEvent"
      responses:
        "202":
          description: Event accepted
          content:
            application/json:
              schema:
                type: object
                required: [accepted]
                properties:
                  accepted:
                    type: boolean
                    example: true
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"

  # ---------------------------------------------------------------------------
  # Session
  # ---------------------------------------------------------------------------
  /v1/auth/session:
    get:
      operationId: getSession
      tags: [Session]
      summary: Get current session
      description: Returns the authenticated user's identity and workspace memberships.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: Current user session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"

  # ---------------------------------------------------------------------------
  # CLI Auth
  # ---------------------------------------------------------------------------
...
  /v1/cli-auth/device/approve:
    post:
      operationId: approveCliDeviceCode
      tags: [CLIAuth]
      summary: Approve a pending CLI verification login
      description: >
        Approves a pending CLI login from the authenticated web session and issues
        the corresponding CLI token for the waiting terminal.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveDeviceCodeRequest"
      responses:
        "200":
          description: Device verification approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceApprovalResponse"
        "400":
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /v1/cli-auth/device/deny:
    post:
      operationId: denyCliDeviceCode
      tags: [CLIAuth]
      summary: Deny a pending CLI verification login
      description: >
        Denies a pending CLI login from the authenticated web session. The
        waiting terminal receives an `access_denied` polling response.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DenyDeviceCodeRequest"
      responses:
        "200":
          description: Device verification denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceApprovalResponse"
        "400":
          description: Invalid or expired verification code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
  /v1/cli-auth/tokens:
    get:
      operationId: listCliTokens
      tags: [CLIAuth]
      summary: List CLI tokens
      description: Lists active CLI tokens owned by the authenticated user.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: Active CLI tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListCLITokensResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createCliToken
      tags: [CLIAuth]
      summary: Create a CLI token
      description: Creates a long-lived CLI token and returns the raw token exactly once.
      security:
        - bearerJWT: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCLITokenRequest"
      responses:
        "201":
          description: CLI token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCLITokenResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/cli-auth/tokens/{id}:
    delete:
      operationId: revokeCliToken
      tags: [CLIAuth]
      summary: Revoke a CLI token
      security:
        - bearerJWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: CLI token revoked
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          $ref: "#/components/responses/NotFoundError"

  # ---------------------------------------------------------------------------
  # Workspace
  # ---------------------------------------------------------------------------
  /v1/workspaces/{workspaceID}/auth-check:
    get:
      operationId: checkWorkspaceAccess
      tags: [Workspace]
      summary: Verify workspace access
      description: Returns success if the caller has access to the specified workspace.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: Workspace access confirmed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceAccessCheckResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"

  # ---------------------------------------------------------------------------
  # User Profile
  # ---------------------------------------------------------------------------
  /v1/users/me:
    get:
      operationId: getUserProfile
      tags: [UserProfile]
      summary: Get current user profile
      description: >
        Returns the authenticated user's profile with full org/workspace
        hierarchy including roles at each level.
      security:
        - bearerJWT: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfileResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Onboarding
  # ---------------------------------------------------------------------------
  /v1/onboarding:
    post:
      operationId: onboard
      tags: [Onboarding]
      summary: First-time onboarding
      description: >
        Creates a new organization and workspace for a first-time user.
        Fails if the user already has an organization.
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardRequest"
      responses:
        "201":
          description: Onboarding complete
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"

  # ---------------------------------------------------------------------------
  # Organizations
  # ---------------------------------------------------------------------------
  /v1/organizations:
    get:
      operationId: listOrganizations
      tags: [Organizations]
      summary: List organizations
      description: Lists organizations the caller belongs to.
      security:
        - bearerJWT: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            default: 50
            minimum: 1
            maximum: 100
        - name: offset
          in: query
          schema:
            type: integer
            format: int32
            default: 0
            minimum: 0
      responses:
        "200":
          description: Paginated list of organizations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListOrganizationsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: createOrganization
      tags: [Organizations]
      summary: Create an organization
      security:
        - bearerJWT: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrganizationResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "409":
          $ref: "#/components/responses/ConflictError"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/workspaces/{workspaceID}/secrets:
    get:
      operationId: listWorkspaceSecrets
      tags: [WorkspaceSecrets]
      summary: List workspace secret keys
      description: Returns secret metadata only — values are never exposed.
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/WorkspaceID"
      responses:
        "200":
          description: List of secret keys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListWorkspaceSecretsResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v1/runtime-profiles/{profileID}:
    get:
      operationId: getRuntimeProfile
      tags: [RuntimeProfiles]
      summary: Get a runtime profile
      security:
        - bearerJWT: []
      parameters:
        - $ref: "#/components/parameters/ProfileID"
      responses:
        "200":
          description: Runtime profile details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RuntimeProfileResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
...
  securitySchemes:
    bearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Bearer token for authenticated API requests. Web sessions use a WorkOS JWT
        and CLI sessions use an AgentClash token prefixed with `clitok_`.
        In development mode, use custom headers instead:
        X-Agentclash-User-Id, X-Agentclash-WorkOS-User-Id,
        X-Agentclash-User-Email, X-Agentclash-User-Display-Name,
        X-Agentclash-Workspace-Memberships.
    hostedRunToken:
      type: http
      scheme: bearer
      description: >
        HMAC-SHA256 callback token issued when a hosted run is started.
        Not a JWT — the token encodes run_id and run_agent_id claims.
    artifactSignature:
      type: apiKey
      in: query
      name: signature
      description: >
        HMAC-SHA256 signature for artifact content URLs. Must be paired with
        an "expires" query parameter (unix timestamp).

  # ---------------------------------------------------------------------------
  # Parameters
  # ---------------------------------------------------------------------------
  parameters:
...
    ArtifactID:
      name: artifactID
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...